requirements.txt
requests==2.28.1  # Para fazer requisições HTTP
Pillow==9.1.1     # Para lidar com imagens no Tkinter
